requirements.txt
# pip install -r requirements.txt

darwin-py==0.8.58
roboflow==1.1.21